requirements.txt
discord
python-dotenv
regex
asyncio
